Added support embedding the native git libraries. Also, added support for the extractor to embed all the .net libraries.

+++ src/SourceIndexingSharp.Extractor/Actions.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using PowerArgs;
+using SourceIndexingSharp.Git;
 
 namespace SourceIndexingSharp.Extractor
 {
@@ -14,14 +15,24 @@ namespace SourceIndexingSharp.Extractor
         {
             ValidateVersion(args);
 
+            if(string.IsNullOrEmpty(args.OutputFileLocation))
+                throw new Exception("You must specify a file location");
+
             if(File.Exists(args.OutputFileLocation))
                 File.Delete(args.OutputFileLocation);
+
             File.WriteAllText(args.OutputFileLocation, "TESTFILE: " + args.OutputFileLocation.ToLower());
         }
 
         private void ValidateVersion(BaseArgs args)
         {
-            var currentVersion = typeof (Actions).Assembly.GetName().Version;
+            var actionsAssemlly = typeof (Actions).Assembly;
+            var currentVersion = actionsAssemlly.GetName().Version;
+            Console.WriteLine("Args version = null ? " + (args.Version == null));
+            Console.WriteLine("Args == null ? " + (args == null));
+            Console.WriteLine("Current version == null ? " + (currentVersion == null));
+            Console.WriteLine("currentVersion:" + currentVersion);
+            Console.WriteLine("version: " + args.Version);
             if(currentVersion < args.Version)
                 throw new SourceIndexException(string.Format("You asking the extractor to run at version {0}, but it was compiled at {1}.", args.Version, currentVersion));
         }
